refactor(dockview): extract stableKeyForTab to shared utility
Deduplicate identical stableKeyForTab function that was copy-pasted
in use-dockview-adapter.ts and use-dockview-persistence.ts.
